# swagger.yaml
openapi: '3.0.3'
info:
  title: Escopo - Trabalho de Conclusão de Curso SENAI
  version: '1.0'
  description: Documentação das rotas da API do projeto Escopo.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Erro:
      type: object
      properties:
        codigo:
          type: integer
        status:
          type: string
        mensagem:
          type: string
        rota:
          type: string
        data_hora:
          type: string
    ErroCampos:
      type: object
      properties:
        codigo:
          type: integer
        status:
          type: string
        erros:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
        rota:
          type: string
        data_hora:
          type: string
...
    Reuniao:
      type: object
      properties:
        titulo:
          type: string
          example: Reunião de Planejamento
        criado_em:
          type: string
          format: date-time
          example: '2026-04-14T10:00:00Z'
        foto_usuarios:
          type: array
          items:
            type: string
          example: ['https://example.com/image1.jpg', 'https://example.com/image2.jpg']
    ReuniaoResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: Reunião de Planejamento
    Link:
      type: object
      properties:
        id:
          type: integer
          example: 1
        url:
          type: string
          example: https://example.com/link
        nome:
          type: string
          example: Link da Reunião
        tipo_link:
          type: string
          example: reuniao
...
    Convidado:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: João Silva
        cargo:
          type: string
          example: Desenvolvedor
    UsuarioReuniao:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Nathan
        cargo:
          type: string
          example: Gerente
        foto_perfil:
          type: string
          example: https://example.com/image.jpg
    ReuniaoDetail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: Reunião de Planejamento
        criado_em:
          type: string
          format: date-time
          example: '2026-04-14T10:00:00Z'
        projeto_id:
          type: integer
          example: 1
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        convidados:
          type: array
          items:
            $ref: '#/components/schemas/Convidado'
        usuarios:
          type: array
          items:
            $ref: '#/components/schemas/UsuarioReuniao'
...
paths:
...
  /api/v1/projeto/{id}/reunioes:
    get:
      tags:
        - Reunião
      summary: Lista reuniões de um projeto específico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reuniao'
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto com o ID informado não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/projeto/{id}/reuniao:
    post:
      tags:
        - Reunião
      summary: Cria uma reunião em um projeto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - titulo
              properties:
                titulo:
                  type: string
                  example: Reunião de Planejamento
                  minLength: 1
                  maxLength: 150
      responses:
        '201':
          description: Reunião criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReuniaoResponse'
              examples:
                criado:
                  summary: Exemplo de resposta
                  value:
                    id: 1
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto com o ID informado não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Campos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/reuniao/{id}/titulo:
    patch:
      tags:
        - Reunião
      summary: Atualiza o título de uma reunião
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID da reunião
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - titulo
              properties:
                titulo:
                  type: string
                  example: Reunião de Planejamento
                  minLength: 1
                  maxLength: 150
      responses:
        '204':
          description: Título atualizado
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Reunião não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Campos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/reuniao/{id}/transcricao:
    patch:
      tags:
        - Reunião
      summary: Atualiza a transcrição de uma reunião
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID da reunião
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - transcricao
              properties:
                transcricao:
                  type: string
                  example: Transcrição da reunião
                  minLength: 1
                  maxLength: 10000
      responses:
        '204':
          description: Transcrição atualizada
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Reunião não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Campos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/reuniao/{id}:
    get:
      tags:
        - Reunião
      summary: Retorna detalhes de uma reunião específica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID da reunião
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReuniaoDetail'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Reunião não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    delete:
      tags:
        - Reunião
      summary: Exclui uma reunião
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID da reunião
      responses:
        '204':
          description: Reunião excluída
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Reunião não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...